backend/requirements.txt
fastapi>=0.104.1,<0.105.0
uvicorn>=0.24.0,<0.25.0
python-multipart>=0.0.6,<0.1.0
pandas>=2.1.3,<2.2.0
openpyxl>=3.1.2,<3.2.0
python-dotenv>=1.0.0,<1.1.0
openai>=1.3.0,<1.4.0 